CA-296489: Allow boot ordering to be set from XS
This uses the CMOS boot order to change boot order. Within each type of boot device (disk/cd/nic) - UEFI boot order is preserved. (So first disk will stay first disk)
Disabling a device moves it to the end of the list, after the UEFI shell, so it'll not boot - as that always succeeds.