CP-28675: XenVariable: Move useful functions into a header file
These functions need to be called from the XenVariable DXE module and
the core image loading infrastructure. Since they are needed for the
image loading infrastructure, they can't be placed in their own library
so instead make them static inline functions. Most of the functions
should be inlined anyway so this is not an issue. Also, the PCI related
functions will soon be removed.