CP-30787: disable unneeded crypto code
Disable EC and DSA in OpenSSL (not needed for secureboot, and avoids ECDSA and DSA X509 public keys).
DH is not used either, but it openssl fails to build if we disable it.

BaseCryptLib is not granular enough, so patch the .inf file to remove the algorithms we don't use.
As far as I can tell this is actually most dead code, everything still links.
Do not remove the files themselves to make rebasing with new upstreams easier.

This should leave just RSA and DH public key types for X509, and RSA signature verification.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>