CA-322248: Prevent guest attempting to accessing priv regsiters
After initialisation, OVMF attempt to clear the screen, but using a
function which zero's the first 0x400000 bytes of the bar.
For QEMU's VGA emulation, this is OK, but for NVIDIA vgpu
(And I presume for gvt-g also), this can end up zeroing beyond
the end of the frame buffer, which upsets the backend.
For NVIDIA, this manifests itself with the NVIDIA drivers
failing to load, and log message on the host report the guest is
attempting to access priv registers.

There does not seem to be any need to clear the screen in this case,
and so the simplest solution is to disable calling this function for
NVIDIA's vgpus.